+++ wiog_include/wiog_data.h
@@ -102,8 +102,8 @@ typedef struct __attribute__((packed)) {
 
 
 
-extern void add_entry_I32 (payload_t* pl, uint8_t type, uint8_t ix, uint32_t st, int32_t val);
-extern void add_entry_I64 (payload_t* pl, uint8_t type, uint8_t ix, uint32_t st, int64_t val);
+extern void add_entry_I32 (payload_t* pl, uint8_t type, uint8_t ix, uint8_t st, int32_t val);
+extern void add_entry_I64 (payload_t* pl, uint8_t type, uint8_t ix, uint8_t st, int64_t val);
 extern void add_entry_buf (payload_t* pl, uint8_t type, uint8_t ix, uint8_t len, uint8_t* buf);
 extern void add_entry_str (payload_t* pl, uint8_t type, uint8_t ix, char* str);
 extern void add_entry_snr (payload_t* pl, dev_uid_t uid, uint8_t snr);
